+++ x/gov/types/genesis.go
@@ -32,8 +32,8 @@ func DefaultGenesis() *GenesisState {
 			MinTxFee:                    100,
 			MaxTxFee:                    1000000,
 			VoteQuorum:                  33,
-			ProposalEndTime:             1, // 1min
-			ProposalEnactmentTime:       2, // 2min
+			ProposalEndTime:             600, // 600 seconds / 10 mins
+			ProposalEnactmentTime:       300, // 300 seconds / 5 mins
 			EnableForeignFeePayments:    true,
 			MischanceRankDecreaseAmount: 10,
 			InactiveRankDecreasePercent: 50,      // 50%