+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [v0.1.17] - 02.16.2021
+### Fixed
+- Problem with ClaimValidator and PubKey encoding due to protocol buff Any type.
+- Fix bug that made that you can vote when the proposal ended.
+- When a proposal does not reach quorum it ends being Quorum not reached.
+- Proposal voting time and enactment time now are defined in seconds.
+- It shows the votes that a proposal has on client query.
+
 ## [v0.1.16.2a] - 02.08.2021
 ### Added
 - Custom evidence module to jail a double signed validator